requirements.txt
async-generator==1.10
attrs==22.1.0
Authlib==1.1.0
backports.cached-property==1.0.2
beautifulsoup4==4.11.1
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docutils==0.19
et-xmlfile==1.1.0
exchangelib==4.7.6
fonttools==4.37.3
fpdf2==2.5.7
furl==2.1.3
graphviz==0.13.2
grpcio==1.49.1
grpcio-tools==1.48.1
h11==0.14.0
htmldocx==0.0.6
hubspot-api-client==4.0.6
idna==3.4
isodate==0.6.1
java-access-bridge-wrapper==0.9.5
jsonpath-ng==1.5.3
jsonschema==4.16.0
lxml==4.9.1
mss==6.1.0
netsuitesdk==1.24.0
notifiers==1.3.3
ntlm-auth==1.5.0
oauthlib==3.2.1
openpyxl==3.0.10
orderedmultidict==1.0.1
outcome==1.2.0
overrides==6.2.0
pdfminer.six==20201018
Pillow==9.2.0
platformdirs==2.5.2
ply==3.11
protobuf==3.20.1
proxy-tools==0.1.0
pycparser==2.21
Pygments==2.13.0
pynput-robocorp-fork==4.0.0
pyotp==2.7.0
PyPDF2==1.28.6
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
python-xlib==0.31
pytz==2022.2.1
pywebview==3.6.3
PyYAML==5.4.1
requests==2.28.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
robocorp-dialog==0.5.3
robotframework==5.0.1
robotframework-assertion-engine==0.5.1
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.3
robotframework-seleniumlibrary==6.0.0
robotframework-seleniumtestability==2.0.0
rpaframework-core==10.0.0
rpaframework-dialogs==4.0.0
rpaframework-pdf==5.0.0
schedule==1.1.0
selenium==4.4.3
simple-salesforce==1.12.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
svg.path==6.2
tenacity==8.1.0
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tweepy==3.10.0
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==2.1
urllib3==1.26.12
webdriver-manager==3.8.3
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.1.0
